+++ src/components/UiTable/index.js
@@ -1,0 +1,4 @@
+import { factoryComponentPlugin } from '@/utils/helpers';
+import UiTable from './UiTable';
+
+export default factoryComponentPlugin(UiTable);