+++ src/components/UiTableCell/index.js
@@ -1,0 +1,4 @@
+import { factoryComponentPlugin } from '@/utils/helpers';
+import UiTableCell from './UiTableCell';
+
+export default factoryComponentPlugin(UiTableCell);