+++ src/components/UiTableColumn/UiTableColumn.js
@@ -1,0 +1,119 @@
+import uiIcon from '../uiIcon';
+
+export default {
+  name: 'ui-table-column',
+  props: {
+    /**
+     * Content alignment inside wrapper
+     */
+    align: {
+      type: String,
+      default: null,
+      validator: v => ['left', 'right', 'center', null].includes(v),
+    },
+    /**
+     * Shows sort icon and make column sortable
+     */
+    sortable: {
+      type: Boolean,
+      default: false,
+    },
+    /**
+     * Sort direction
+     */
+    direction: {
+      type: String,
+      default: null,
+      validator: v => ['asc', 'desc'].includes(v),
+    },
+  },
+  methods: {
+    /**
+     * Click event handler
+     *
+     * @param {MouseEvent} e
+     */
+    clickHandler(e) {
+      this.$emit('click', e);
+
+      if (this.sortable) {
+        const v = this.direction === null ? 'desc' : this.direction === 'desc' ? 'asc' : null;
+
+        this.$emit('toggle', v);
+      }
+    },
+    /**
+     * Generates column template
+     *
+     * @param {Array<VNode>} content
+     * @return {VNode}
+     */
+    genColumn(content) {
+      return this.$createElement(
+        'th',
+        {
+          staticClass: 'table-col',
+          class: {
+            '-sortable': this.sortable,
+          },
+          on: {
+            click: e => this.clickHandler(e),
+          },
+        },
+        content,
+      );
+    },
+    /**
+     * Generates title's wrapper template
+     *
+     * @param {Array<VNode>} content
+     * @return {VNode}
+     */
+    genWrapper(content) {
+      return this.$createElement(
+        'span',
+        {
+          staticClass: 'table-col__content',
+          class: {
+            '-right': this.align === 'right',
+            '-center': this.align === 'center',
+          },
+        },
+        content,
+      );
+    },
+    /**
+     * Generates title template
+     *
+     * @return {VNode}
+     */
+    genTitle() {
+      return this.$createElement('span', { staticClass: 'table-col__title' }, this.$slots.default);
+    },
+    /**
+     * Generates sort icon template
+     *
+     * @return {VNode | null}
+     */
+    genIcon() {
+      if (!this.sortable) {
+        return null;
+      }
+
+      return this.$createElement(uiIcon, {
+        staticClass: 'table-col__icon',
+        class: {
+          '-asc': this.direction === 'asc',
+          '-desc': this.direction === 'desc',
+        },
+        props: {
+          name: 'sort',
+          interactive: true,
+        },
+      });
+    },
+  },
+  render() {
+    return this.genColumn([this.genWrapper([this.genTitle(), this.genIcon()])]);
+  },
+};