+++ src/components/UiTableColumn/index.js
@@ -1,0 +1,4 @@
+import { factoryComponentPlugin } from '@/utils/helpers';
+import UiTableColumn from './UiTableColumn';
+
+export default factoryComponentPlugin(UiTableColumn);