+++ src/components/UiTableRow/index.js
@@ -1,0 +1,4 @@
+import { factoryComponentPlugin } from '@/utils/helpers';
+import UiTableRow from './UiTableRow';
+
+export default factoryComponentPlugin(UiTableRow);